pyproject.toml
[project]
name = "inventory-management-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "fastapi>=0.116.1",
    "fastapi-pagination>=0.13.3",
    "psycopg2>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "sqlakeyset>=2.0.1746777265",
    "sqlalchemy[asyncio]>=2.0.42",
    "sqlmodel[asyncio]>=0.0.24",
    "uvicorn[standard]>=0.35.0",
]
